+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fixed the bug that "Parameter is not set" error in word modifier
+- Fixed a bug that satisfy matcher succeed even syntax error
 
 ## [0.26.1] - 2020-07-13